+++ src/channels/ScreenChannel.php
@@ -2,9 +2,7 @@
 
 namespace webzop\notifications\channels;
 
-use Yii;
 use webzop\notifications\Channel;
-use webzop\notifications\Notification;
 
 class ScreenChannel extends Channel
 {
@@ -14,7 +12,6 @@ class ScreenChannel extends Channel
     public function sendNotification($notification)
     {
         // The notification will be shown via JS
-        return TRUE;
+        return true;
     }
-
 }